Separated model and controller.
However, more work needs to be done because the controller should not
have direct access to the viewports. The direct pointers to BRDFViewport
and ModelViewport should be handled by a separate main.js.
