+++ Scripts/Model.js
@@ -1,8 +1,7 @@
 "use strict";
 
-import {loadBRDF_disneyFormat, compile_and_link_shdr,
-  loadBrdfFile} from './gl-wrangling-funcs.js';
-import KnobInput from "./KnobInput.js";
+import {loadBRDF_disneyFormat, compile_and_link_shdr} from './gl-wrangling-funcs.js';
+import {map_insert_chain} from './collections-wranglers.js';
 
 //requires d3.js
 
@@ -16,245 +15,272 @@ import KnobInput from "./KnobInput.js";
 //Currently there are no "constructor" arguments
 //See BRDFViewport.js for an example of one with arguments
 export default function Model(){
-  const starting_theta = 45;
-  const starting_phi = 0;
+  //const starting_theta = 45;
+  //const starting_phi = 0;
 
   //Declare our object's properties and methods below.
   //They are private by default, unless we put them
   //in the "frozen" object that gets returned at the end.
   let
   viewers = [],
-  incidentThetaEnvelope,
-  incidentPhiEnvelope,
+  //incidentThetaEnvelope,
+  //incidentPhiEnvelope,
   //normalThetaEnvelope,
   //normalPhiEnvelope,
   registerViewer = function(new_viewer){
     viewers.push(new_viewer);
   },
-  setupUI = function() {
-    //FIXME: ControlsManager should be writing to its
-    //own div, not to #brdf-menu
-    let menu = d3.select("#brdf-menu");
-    let thetaInput;
-    let thetaOutput;
-    let phiInput;
-    let phiOutput;
-    let camRotInput;
 
-    menu.html("");
-
-    let fileChooser = menu.append("div")
-    .attr("id", "file-chooser")
-    .style("display", "flex");
-
-    menu.append("br");
-
-    fileChooser.html("");
-    fileChooser.append("input")
-    .attr("id", "file_chooser")
-    .attr("type","file");
-
-    let sliderDiv = menu.append("div");
-    sliderDiv.style("display", "flex")
-    .style("width", "100%");
-
-    /* Add incident theta slider */
-    incidentThetaEnvelope = addEnvelopeControl(sliderDiv, "θ",
-      "slider_incidentTheta", 0, 90, starting_theta);
-
-    /* Add incident phi slider */
-    incidentPhiEnvelope = addEnvelopeControl(sliderDiv, "φ",
-      "slider_incidentPhi", -180, 180, starting_phi);
-
-    /* add camRot slider */
-    d3.select("#brdf-header").html("");
-    d3.select("#brdf-header").append("input")
-    .attr("id", "slider_camRot")
-    .attr("class", "niceSlider")
-    .attr("type", "range")
-    .attr("min", -180)
-    .attr("max", 180)
-    .attr("step", 1)
-    .attr("value", 0);
-  },
-
-  getSupportedPropertyName = function(properties) {
-    for (var i = 0; i < properties.length; i++)
-      if (typeof document.body.style[properties[i]] !== 'undefined')
-        return properties[i];
-      return null;
-  },
-
-  getTransformProperty = function() {
-    return getSupportedPropertyName([
-      'transform', 'msTransform', 'webkitTransform', 'mozTransform', 'oTransform'
-      ]);
-  },
-
-  debounce = function(func, wait, immediate) {
-    var timeout;
-    return function() {
-      var context = this, args = arguments;
-      var later = function() {
-        timeout = null;
-        if (!immediate) func.apply(context, args);
-      };
-      var callNow = immediate && !timeout;
-      clearTimeout(timeout);
-      timeout = setTimeout(later, wait);
-      if (callNow) func.apply(context, args);
-    };
-  },
-
-  addEnvelopeControl = function(menu, name, rangeId, minimum, maximum, initial_value) {
-    let scale = 1.0;
-
-    let control = menu.append("div")
-    .attr("class", "fl-studio-envelope__control");
-
-    let knob = control.append("div")
-    .attr("class", "knob-input fl-studio-envelope__knob envelope-knob " + name);
-
-    let visual = knob.append("svg")
-    .attr("class", "knob-input__visual")
-    .attr("viewBox", "0 0 " + 40 * (1.0/scale) + " " + 40 * (1.0/scale) );
-
-    let focus_indicator = visual.append("circle")
-    .attr("class", "focus-indicator")
-    .attr("cx", 20).attr("cy", 20).attr("r", 18)
-    .attr("fill", "#ffffff")
-    .attr("filter", "url(#glow)");
-
-    let indicator_ring_bg = visual.append("circle")
-    .attr("class", "indicator-ring-bg")
-    .attr("cx", 20).attr("cy", 20).attr("r", 18)
-    .attr("fill", "#353b3f").attr("stroke", "#23292d");
-
-    let indicator_ring = visual.append("path")
-    .attr("class", "indicator-ring")
-    .attr("d", "M20,20Z")
-    .attr("fill", "#ffffff");
-
-    let dial = visual.append("g").attr("class", "dial");
-
-    dial.append("circle")
-    .attr("cx", 20).attr("cy", 20)
-    .attr("r", 16).attr("fill", "url(#grad-dial-soft-shadow)");
-
-    dial.append("ellipse")
-    .attr("cx", 20).attr("cy", 22).attr("rx", 14).attr("ry", 14.5)
-    .attr("fill", "#242a2e").attr("opacity", 0.15);
-
-    dial.append("circle")
-    .attr("cx", 20).attr("cy", 20).attr("r", 14)
-    .attr("fill", "url(#grad-dial-base)")
-    .attr("stroke", "#242a2e")
-    .attr("stroke-width", 1.5);
-
-    dial.append("circle")
-    .attr("cx", 20).attr("cy", 20).attr("r", 13)
-    .attr("fill", "transparent")
-    .attr("stroke", "url(#grad-dial-highlight)")
-    .attr("stroke-width", 1.5);
-
-    // let dial_highlight = dial.append("circle")
-    //   .attr("class", "dial-highlight")
-    //   .attr("cx", 20).attr("cy", 20).attr("r", 14)
-    //   .attr("fill", "#ffffff");
-
-    let indicator_dot = dial.append("circle")
-    .attr("class", "indicator-dot")
-    .attr("cx", 20).attr("cy", 30).attr("r", 1.5)
-    .attr("fill", "#ffffff");
-
-    let label = control.append("div")
-    .attr("class", "fl-studio-envelope__label")
-    .text(name);
-
-    let transformProp = getTransformProperty();
-
-    return new KnobInput(knob.node(), rangeId, {
-      visualContext: function() {
-        this.indicatorRing = this.element.querySelector('.indicator-ring');
-        var ringStyle = getComputedStyle(this.element.querySelector('.indicator-ring-bg'));
-        this.r = parseFloat(ringStyle.r) - (parseFloat(ringStyle.strokeWidth) / 2);
-        this.indicatorDot = this.element.querySelector('.indicator-dot');
-        this.indicatorDot.style[`${transformProp}Origin`] = '20px 20px';
-      },
-      updateVisuals: function(norm) {
-        var theta = Math.PI*2*norm + 0.5*Math.PI;
-        var endX = this.r*Math.cos(theta) + 20;
-        var endY = this.r*Math.sin(theta) + 20;
-        // using 2 arcs rather than flags since one arc collapses if it gets near 360deg
-        this.indicatorRing.setAttribute('d',`M20,20l0,${this.r}${norm> 0.5?`A${this.r},${this.r},0,0,1,20,${20-this.r}`:''}A-${this.r},${this.r},0,0,1,${endX},${endY}Z`);
-        this.indicatorDot.style[transformProp] = `rotate(${360*norm}deg)`;
-      },
-      min: minimum,
-      max: maximum,
-      initial: initial_value,
+  setTheta = function(newTheta){
+    viewers.forEach(function(v) {
+      if (v.inputByModel === true) {
+        v.updateTheta(newTheta);
+      }
     });
   },
 
-  setupUICallbacks = function() {
-    //Set initial values
-    //now this slider only controls light theta and phi
-    incidentThetaEnvelope.addEventListener('change', (event) => {
-      viewers.forEach(function(v) {
-        let new_theta = incidentThetaEnvelope.value;
-
-         viewers.forEach(function(v) {
-           let new_theta = event.target.value;
-           if (v.inputByModel === true) {
-             v.updateTheta(new_theta);
-           }
-         });
-      });
+  setPhi = function(newPhi){
+    viewers.forEach(function(v) {
+      if (v.inputByModel === true) {
+        v.updatePhi(newPhi);
+      }
     });
+  },
 
-    incidentPhiEnvelope.addEventListener('change', (event) => {
-      let new_phi = incidentPhiEnvelope.value;
-
-       viewers.forEach(function(v) {
-         let new_phi = event.target.value;
-         if (v.inputByModel === true) {
-           v.updatePhi(new_phi);
-         }
-       });
+  setCamRot = function(newCamRot){
+    viewers.forEach(function(v) {
+      if (v.inputByModel === true && "updateCamRot" in v) {
+        v.updateCamRot(newCamRot);
+      }
     });
+  },
 
-    document.getElementById("slider_camRot").oninput = (event) => {
-      viewers.forEach(function(v) {
-        let new_camRot = event.target.value;
-        if( "updateCamRot" in v ){
-          v.updateCamRot(new_camRot);
-        }
+  loadBrdfFile = function(fileList, viewers){
+    let reader = new FileReader();
+    //key: uniform name. value: function for updating the uniform.
+    let uniform_update_funcs = new Map();
+    let uniforms;
+    let vertSrc;
+    let fragSrc;
+
+    reader.onload = function() {
+      //FIXME: duplicate definition of shdrDir
+      let loadBRDFPromise = loadBRDF_disneyFormat({brdfFileStr: reader.result,
+        shdrDir: "./Shaders/", templatePath: "lobe_template.vert",
+        vertPath: "lobe.vert", fragPath: "phong.frag", templateType: "vert"});
+
+      loadBRDFPromise.then(value => {
+        uniforms = value.uniformsInfo;
+        vertSrc = value.finalVtxSrc;
+        fragSrc = value.finalFragSrc;
+        //console.log("Loading .brdf done!");
+        //console.log(vertSrc);
+        //console.log(fragSrc);
+      }, err => {
+          throw "BRDF load error: " + err;
+      }).then( () => { //call the below asynchronously, AFTER the above is done loading
+        //The below is passed as a first-class function to addUniformsFunc
+        let addUniformsHelper = function(gl) {
+          let program = compile_and_link_shdr(gl, vertSrc, fragSrc);
+
+          gl.useProgram(program);
+
+          Object.keys(uniforms).forEach( u => {
+            //We need to use "let" for loc. If we use "var", the closure
+            //won't work due to hoisting and all functions stored in
+            //uniform_update_funcs will store a pointer to the location of
+            //the last uniform processed!
+            let loc = gl.getUniformLocation(program, u);
+            let curr_u = uniforms[u];
+            let u_type = curr_u.type;
+
+            if (u_type === "float") {
+              let flt_update_func = flt => {
+                gl.uniform1f(loc,flt);
+              };
+
+              if ( (typeof curr_u.default) !== "number" ){
+                throw "curr_u.default should be a number!";
+              }
+              flt_update_func(curr_u.default);
+
+              map_insert_chain(uniform_update_funcs, u, flt_update_func);
+            } else if (u_type === "bool") {
+              let bool_update_func = bool_v => {
+                if (bool_v === true) {
+                  gl.uniform1i(loc,1);
+                } else if (bool_v === false) {
+                  gl.uniform1i(loc,0);
+                } else {
+                  throw "Invalid boolean input: " + bool_v;
+                }
+              };
+
+              bool_update_func(curr_u.default);
+
+              map_insert_chain(uniform_update_funcs, u, bool_update_func);
+            } else if (u_type === "color") {
+              let vec3_update_func = vec3_v => {
+                gl.uniform3f(loc, vec3_v[0], vec3_v[1], vec3_v[2]);
+              };
+
+              if ( (typeof curr_u.defaultR) !== "number" ||
+                   (typeof curr_u.defaultG) !== "number" ||
+                   (typeof curr_u.defaultB) !== "number" ){
+                throw "curr_u.default[RGB] should be a number!";
+              }
+              vec3_update_func([curr_u.defaultR, curr_u.defaultG,
+                curr_u.defaultB]);
+
+              map_insert_chain(uniform_update_funcs, u, vec3_update_func);
+            } else {
+              throw "Invalid uniform type: " + u_type;
+            }
+          });
+
+          //Give our new program to the Viewer, so the Viewer can bind
+          //it to the appropriate variable. E.g. At time of writing,
+          //BRDFViewport binds this to lobeProgram.
+          return program;
+        };
+
+        viewers.forEach( v => {
+          if( "addUniformsFunc" in v ){
+            v.addUniformsFunc(addUniformsHelper);
+          }
+        });
+      }).then( () => {
+        console.log("Done adding uniforms!");
+        //TODO: bind each function in uniform_update_funcs to its own slider.
       });
     };
 
-    //File input snippet from:
-    //https://developer.mozilla.org/en-US/docs/Web/API/File/Using_files_from_web_applications
-    document.getElementById("file_chooser").addEventListener("change",
-      //in the below function, "this" appears to be bound to some object
-      //that addEventListener bind the function to.
-      function(){
-        loadBrdfFile(this.files,viewers);
-      }, false);
+    //onload will be invoked when this is done
+    reader.readAsText(fileList[0]);
+  },
+
+  loadAnalyticalBRDF = function(in_fileList){
+    loadBrdfFile(in_fileList, viewers);
   };
 
+  //setupUI = function() {
+    //let menu = d3.select("#brdf-menu");
+    //let thetaInput;
+    //let thetaOutput;
+    //let phiInput;
+    //let phiOutput;
+    //let camRotInput;
 
-  //TODO: should the "controller" really be processing the BRDF file?
-  //This sounds more like the job of the "model".
-  //Because this is still a small project we probably don't need a separate
-  //"model" and "controller"...
+    //menu.html("");
 
+    //let fileChooser = menu.append("div")
+    //.attr("id", "file-chooser")
+    //.style("display", "flex");
+
+    //menu.append("br");
+
+    //fileChooser.html("");
+    //fileChooser.append("input")
+    //.attr("id", "file_chooser")
+    //.attr("type","file");
+
+    //let sliderDiv = menu.append("div");
+    //sliderDiv.style("display", "flex")
+    //.style("width", "100%");
+
+    //[> Add incident theta slider <]
+    //incidentThetaEnvelope = addEnvelopeControl(sliderDiv, "θ",
+      //"slider_incidentTheta", 0, 90, starting_theta);
+
+    //[> Add incident phi slider <]
+    //incidentPhiEnvelope = addEnvelopeControl(sliderDiv, "φ",
+      //"slider_incidentPhi", -180, 180, starting_phi);
+
+    //[> add camRot slider <]
+    //d3.select("#brdf-header").html("");
+    //d3.select("#brdf-header").append("input")
+    //.attr("id", "slider_camRot")
+    //.attr("class", "niceSlider")
+    //.attr("type", "range")
+    //.attr("min", -180)
+    //.attr("max", 180)
+    //.attr("step", 1)
+    //.attr("value", 0);
+  //},
+
+
+  //debounce = function(func, wait, immediate) {
+    //var timeout;
+    //return function() {
+      //var context = this, args = arguments;
+      //var later = function() {
+        //timeout = null;
+        //if (!immediate) func.apply(context, args);
+      //};
+      //var callNow = immediate && !timeout;
+      //clearTimeout(timeout);
+      //timeout = setTimeout(later, wait);
+      //if (callNow) func.apply(context, args);
+    //};
+  //};
+
+  //setupUICallbacks = function() {
+    ////Set initial values
+    ////now this slider only controls light theta and phi
+    //incidentThetaEnvelope.addEventListener('change', (event) => {
+      //viewers.forEach(function(v) {
+        //let new_theta = incidentThetaEnvelope.value;
+
+         //viewers.forEach(function(v) {
+           //let new_theta = event.target.value;
+           //if (v.inputByModel === true) {
+             //v.updateTheta(new_theta);
+           //}
+         //});
+      //});
+    //});
+
+    //incidentPhiEnvelope.addEventListener('change', (event) => {
+      //let new_phi = incidentPhiEnvelope.value;
+
+       //viewers.forEach(function(v) {
+         //let new_phi = event.target.value;
+         //if (v.inputByModel === true) {
+           //v.updatePhi(new_phi);
+         //}
+       //});
+    //});
+
+    //document.getElementById("slider_camRot").oninput = (event) => {
+      //viewers.forEach(function(v) {
+        //let new_camRot = event.target.value;
+        //if( "updateCamRot" in v ){
+          //v.updateCamRot(new_camRot);
+        //}
+      //});
+    //};
+
+    ////File input snippet from:
+    ////https://developer.mozilla.org/en-US/docs/Web/API/File/Using_files_from_web_applications
+    //document.getElementById("file_chooser").addEventListener("change",
+      ////in the below function, "this" appears to be bound to some object
+      ////that addEventListener bind the function to.
+      //function(){
+        //loadBrdfFile(this.files,viewers);
+      //}, false);
+  //};
 
   //************* Start "constructor" **************
-  setupUI();
-  setupUICallbacks();
+  //setupUI();
+  //setupUICallbacks();
   //************* End "constructor" **************
 
   //Put any methods / properties that we want to make pulic inside this object.
   return Object.freeze({
-    registerViewer
+    registerViewer,
+    setTheta,
+    setPhi,
+    setCamRot,
+    loadAnalyticalBRDF,
   });
 }
