+++ Scripts/collections-wranglers.js
@@ -1,5 +1,3 @@
-"use strict";
-
 //Assumes that map is an ES6 Map or WeakMap.
 //Alternate insertion function such that we have a hashmap WITH CHAINING.
 export function map_insert_chain(map, key, val){
