+++ Scripts/gl-wrangling-funcs.js
@@ -1,110 +1,7 @@
 import {deg2rad, rotY, rotZ} from './math-utils.js';
 import {getNextLine_brdfFile} from './text-utils.js';
-import {map_insert_chain} from './collections-wranglers.js';
+//import {map_insert_chain} from './collections-wranglers.js';
 
-export function loadBrdfFile(fileList, viewers){
-  let reader = new FileReader();
-  //key: uniform name. value: function for updating the uniform.
-  let uniform_update_funcs = new Map();
-  let uniforms;
-  let vertSrc;
-  let fragSrc;
-
-  reader.onload = function() {
-    //FIXME: duplicate definition of shdrDir
-    let loadBRDFPromise = loadBRDF_disneyFormat({brdfFileStr: reader.result,
-      shdrDir: "./Shaders/", templatePath: "lobe_template.vert",
-      vertPath: "lobe.vert", fragPath: "phong.frag", templateType: "vert"});
-
-    loadBRDFPromise.then(value => {
-      uniforms = value.uniformsInfo;
-      vertSrc = value.finalVtxSrc;
-      fragSrc = value.finalFragSrc;
-      //console.log("Loading .brdf done!");
-      //console.log(vertSrc);
-      //console.log(fragSrc);
-    }, err => {
-        throw "BRDF load error: " + err;
-    }).then( () => { //call the below asynchronously, AFTER the above is done loading
-      //The below is passed as a first-class function to addUniformsFunc
-      let addUniformsHelper = function(gl) {
-        let program = compile_and_link_shdr(gl, vertSrc, fragSrc);
-
-        gl.useProgram(program);
-
-        Object.keys(uniforms).forEach( u => {
-          //We need to use "let" for loc. If we use "var", the closure
-          //won't work due to hoisting and all functions stored in
-          //uniform_update_funcs will store a pointer to the location of
-          //the last uniform processed!
-          let loc = gl.getUniformLocation(program, u);
-          let curr_u = uniforms[u];
-          let u_type = curr_u.type;
-
-          if (u_type === "float") {
-            let flt_update_func = flt => {
-              gl.uniform1f(loc,flt);
-            };
-
-            if ( (typeof curr_u.default) !== "number" ){
-              throw "curr_u.default should be a number!";
-            }
-            flt_update_func(curr_u.default);
-
-            map_insert_chain(uniform_update_funcs, u, flt_update_func);
-          } else if (u_type === "bool") {
-            let bool_update_func = bool_v => {
-              if (bool_v === true) {
-                gl.uniform1i(loc,1);
-              } else if (bool_v === false) {
-                gl.uniform1i(loc,0);
-              } else {
-                throw "Invalid boolean input: " + bool_v;
-              }
-            };
-
-            bool_update_func(curr_u.default);
-
-            map_insert_chain(uniform_update_funcs, u, bool_update_func);
-          } else if (u_type === "color") {
-            let vec3_update_func = vec3_v => {
-              gl.uniform3f(loc, vec3_v[0], vec3_v[1], vec3_v[2]);
-            };
-
-            if ( (typeof curr_u.defaultR) !== "number" ||
-                 (typeof curr_u.defaultG) !== "number" ||
-                 (typeof curr_u.defaultB) !== "number" ){
-              throw "curr_u.default[RGB] should be a number!";
-            }
-            vec3_update_func([curr_u.defaultR, curr_u.defaultG,
-              curr_u.defaultB]);
-
-            map_insert_chain(uniform_update_funcs, u, vec3_update_func);
-          } else {
-            throw "Invalid uniform type: " + u_type;
-          }
-        });
-
-        //Give our new program to the Viewer, so the Viewer can bind
-        //it to the appropriate variable. E.g. At time of writing,
-        //BRDFViewport binds this to lobeProgram.
-        return program;
-      };
-
-      viewers.forEach( v => {
-        if( "addUniformsFunc" in v ){
-          v.addUniformsFunc(addUniformsHelper);
-        }
-      });
-    }).then( () => {
-      console.log("Done adding uniforms!");
-      //TODO: bind each function in uniform_update_funcs to its own slider.
-    });
-  };
-
-  //onload will be invoked when this is done
-  reader.readAsText(fileList[0]);
-}
 
 export function loadBRDF_disneyFormat(spec){
   let { brdfFileStr, shdrDir, templatePath, vertPath, fragPath,
